+++ components/FormExpenses.js
@@ -3,6 +3,7 @@ import { View, StyleSheet, ScrollView, KeyboardAvoidingView } from 'react-native
 import { useFormik } from "formik";
 import { Appbar, TextInput, Text, useTheme, ActivityIndicator,} from 'react-native-paper';
 import { useAppContext } from "../providers/AppContextProvider";
+import { useCreateItemContext, useItemContext } from "../providers/ItemContextProvider";
 
 const InputField = (props) => <TextInput keyboardType="numeric"
   style={{ marginTop: 12, marginHorizontal: 2 }}
@@ -14,7 +15,8 @@ const InputField = (props) => <TextInput keyboardType="numeric"
 
 export default function FormExpenses({ route, navigation }) {
 
-  const { settings, appliedSettings } = useAppContext();
+  const { settings } = useAppContext();
+  const { appliedSettings } = useItemContext();
   const [loaded, setLoaded] = useState(false);
   const nameForma = 'Редактирование затрат' 
   //const theme = useTheme()
